Bullets

+++ ProjectGeometry/Bullet.h
@@ -7,13 +7,13 @@ public:
 	enum TypeBullet
 	{
 	};
-	int damageB = 2;
+	float damageB = 2;
 	float scaleB = 0.6f;
 	float radiusB = 10.0f;
 	float speedB = 1000;
 	sf::CircleShape shapeB;
 	sf::Vector2f trajectoire;
-	Bullet(int damage, sf::Vector2f direction);
+	Bullet(float damage, sf::Vector2f direction);
 	void Update(sf::RenderWindow *window, float deltaTime);
 	void MoveBullet(float speed);
 };
